% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnm-functions.R
\name{ddirichlet}
\alias{ddirichlet}
\title{Density of the Dirichlet distribution.}
\usage{
ddirichlet(x, alpha, log = FALSE)
}
\arguments{
\item{x}{Observation vector of nonnegative entries which sum to one.}

\item{alpha}{Weight parameter: a vector of nonnegative entries the same length as \code{x}.}

\item{log}{Logical; whether to evaluate the density on the log scale.}
}
\value{
The density or log-density evaluated at the inputs (scalar).
}
\description{
Density of the Dirichlet distribution.
}
