% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnm-gibbs-sampler.R
\name{gibbs_sampler}
\alias{gibbs_sampler}
\title{Formal Gibbs Sampler for all five parameters of normal-normal-mixture (NNM) model for clustering}
\usage{
gibbs_sampler(
  nsamples,
  burn,
  K,
  Y,
  V,
  theta_init,
  mu_init,
  Sigma_init,
  Z_init,
  rho_init,
  return_Z = FALSE,
  return_theta = FALSE
)
}
\arguments{
\item{nsamples}{An integer represents the number of samples the Gibbs Sampler will perform}

\item{burn}{An integer represents the burn-in sampler of the Gibbs Sampler}

\item{K}{An integer represents the number of mixture components.}

\item{Y}{An \verb{N x p} matrix of observations, each of which is a column.}

\item{V}{A \verb{p x p x N} matrix of variances corresponding to each observation.}

\item{theta_init}{A \verb{N x p} matrix of observation means.}

\item{mu_init}{A \verb{K x p} matrix of group means, where \code{K} is the number of mixture components and \code{p} is the dimension of each observation.}

\item{Sigma_init}{A \verb{p x p x K} matrix of group variances.}

\item{Z_init}{A length-\code{N} vector of integers between 1 and \code{K} indicating the group membership of each column of \code{theta}.}

\item{rho_init}{A length-\code{K} probability vector of group membership.}

\item{return_Z}{An True/False parameter indicates whether to store z during every sample process or not}

\item{return_theta}{An True/False parameter indicates whether to store theta during every sample process or not}
}
\value{
A list contains all updated parameters
}
\description{
Formal Gibbs Sampler for all five parameters of normal-normal-mixture (NNM) model for clustering
}
