% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnm-functions.R
\name{init_z}
\alias{init_z}
\title{Initial cluster allocation.}
\usage{
init_z(y, K, max_iter = 10, nstart = 10)
}
\arguments{
\item{y}{An \verb{N x p} matrix of observations, each of which is a column.}

\item{K}{Number of clusters (positive integer).}

\item{max_iter}{The maximum number of steps in the \code{\link[stats:kmeans]{stats::kmeans()}} algorithm.}

\item{nstart}{The number of random starts in the \code{\link[stats:kmeans]{stats::kmeans()}} algorithm.}
}
\value{
A vector of length \code{N} consisting of integers between \code{1} and \code{K} specifying an initial clustering of the observations.
}
\description{
Initializes the clusters using the kmeans++ algorithm of Arthur & Vassilvitskii (2007).
}
\references{
Arthur, D., Vassilvitskii, S. "k-means++: the advantages of careful seeding" \emph{Proceedings of the 18th Annual ACM-SIAM Symposium on Discrete Algorithms.} (2007): 1027–1035. \url{http://ilpubs.stanford.edu:8090/778/1/2006-13.pdf}.
}
