% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnm-functions.R
\name{nnm_rest}
\alias{nnm_rest}
\title{Prior data loglikelihood for the normal-normal-mixture (NNM) model.}
\usage{
nnm_rest(K, p, vk, omega, Sigma)
}
\arguments{
\item{K}{An integer represents the number of mixture components.}

\item{p}{An integer represents is the dimension of each observation.}

\item{vk}{An integer used in Inverse-Wishart distribution which set to \code{p + 2}.}

\item{omega}{An \verb{p x p} symmetric positive-definite matrix used in Inverse-Wishart distribution.}

\item{Sigma}{A \verb{p x p x K} matrix of group variances.}
}
\value{
The Prior data loglikelihood evaluated at the inputs (scalar).
}
\description{
Prior data loglikelihood for the normal-normal-mixture (NNM) model.
}
