% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnmGibbs-package.R
\docType{package}
\name{nnmGibbs}
\alias{nnmGibbs}
\title{Inference for the MCMC Gibbs Sampler of the NNM Model}
\details{
Provide the the Conditional Update Distrubution Function of each parameter from the NNM Model,
and the MCMC Gibbs Sampler of the NNM Model, which require package mniw
}
\keyword{"_PACKAGE"}
\keyword{internal}
