% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnm-functions.R
\name{rcategorical}
\alias{rcategorical}
\title{Sample from a categorical distribution.}
\usage{
rcategorical(prob)
}
\arguments{
\item{prob}{An \verb{n_cat x n_prob} matrix of probability vectors, each of which is a column of \code{prob}.  The entries of \code{prob} must be nonnegative, but are internally normalized such that each column sums to one.}
}
\value{
A vector of length \code{n_prob} of integers between 1 and \code{n_cat} indicating the category draw for each column of \code{prob}.
}
\description{
Performs one draw from a categorical distribution (i.e., a multinomial distribution with size \code{n = 1}) for each of multiple probability vectors.
}
