% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnm-functions.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Random sampling from the Dirichlet distribution.}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{Number of random draws.}

\item{alpha}{Weight parameter: a vector of nonnegative entries.}
}
\value{
A matrix of size \verb{n x length(alpha)} of which each row is a random draw.
}
\description{
Random sampling from the Dirichlet distribution.
}
