% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnm-gibbs-sampler.R
\name{rho.loglik}
\alias{rho.loglik}
\title{Conditional Distrubution of rho given other parameters}
\usage{
rho.loglik(rho, mu, z)
}
\arguments{
\item{rho}{A length-\code{K} probability vector of group membership.}

\item{mu}{A \verb{K x p} matrix of group means, where \code{K} is the number of mixture components and \code{p} is the dimension of each observation.}

\item{z}{A length-\code{N} vector of integers between 1 and \code{K} indicating the group membership of each column of \code{theta}.}
}
\value{
The conditional log-PDF of rho given other parameters
}
\description{
Conditional Distrubution of rho given other parameters
}
