% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnm-gibbs-sampler.R
\name{theta.loglik}
\alias{theta.loglik}
\title{Conditional Distrubution of theta given other parameters}
\usage{
theta.loglik(theta, mu, Sigma, z, V, y)
}
\arguments{
\item{theta}{A \verb{N x p} matrix of observation means.}

\item{mu}{A \verb{K x p} matrix of group means, where \code{K} is the number of mixture components and \code{p} is the dimension of each observation.}

\item{Sigma}{A \verb{p x p x K} matrix of group variances.}

\item{z}{A length-\code{N} vector of integers between 1 and \code{K} indicating the group membership of each column of \code{theta}.}

\item{V}{A \verb{p x p x N} matrix of variances corresponding to each observation.}

\item{y}{An \verb{N x p} matrix of observations, each of which is a column.}
}
\value{
The conditional log-PDF of theta given other parameters
}
\description{
Conditional Distrubution of theta given other parameters
}
