% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnm-gibbs-sampler.R
\name{update_rho}
\alias{update_rho}
\title{Conditional Draw of rho given other parameters}
\usage{
update_rho(mu_curr, Z_curr)
}
\arguments{
\item{mu_curr}{A \verb{K x p} matrix of group means, where \code{K} is the number of mixture components and \code{p} is the dimension of each observation.}

\item{Z_curr}{A length-\code{N} vector of integers between 1 and \code{K} indicating the group membership of each column of \code{theta}.}
}
\value{
The conditional draw of rho given other parameters
}
\description{
Conditional Draw of rho given other parameters
}
