% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnm-gibbs-sampler.R
\name{update_theta}
\alias{update_theta}
\title{Conditional Draw of theta given other parameters}
\usage{
update_theta(mu_curr, Sigma_curr, Z_curr, Y, V)
}
\arguments{
\item{mu_curr}{A \verb{K x p} matrix of group means, where \code{K} is the number of mixture components and \code{p} is the dimension of each observation.}

\item{Sigma_curr}{A \verb{p x p x K} matrix of group variances.}

\item{Z_curr}{A length-\code{N} vector of integers between 1 and \code{K} indicating the group membership of each column of \code{theta}.}

\item{Y}{An \verb{N x p} matrix of observations, each of which is a column.}

\item{V}{A \verb{p x p x N} matrix of variances corresponding to each observation.}
}
\value{
The conditional draw of theta given other parameters
}
\description{
Conditional Draw of theta given other parameters
}
