% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnm-gibbs-sampler.R
\name{update_Z}
\alias{update_Z}
\title{Conditional Draw of z given other parameters}
\usage{
update_Z(theta_curr, mu_curr, Sigma_curr, rho_curr, Y)
}
\arguments{
\item{theta_curr}{A \verb{N x p} matrix of observation means.}

\item{mu_curr}{A \verb{K x p} matrix of group means, where \code{K} is the number of mixture components and \code{p} is the dimension of each observation.}

\item{Sigma_curr}{A \verb{p x p x K} matrix of group variances.}

\item{rho_curr}{A length-\code{K} probability vector of group membership.}

\item{Y}{An \verb{N x p} matrix of observations, each of which is a column.}
}
\value{
The conditional draw of z and lambda given other parameters
}
\description{
Conditional Draw of z given other parameters
}
